Rename SHAFT MCP module to shaft-mcp
