+++ .github/workflows/shaft-mcp.yml
@@ -1,4 +1,4 @@
-name: SHAFT MCP
+name: shaft-mcp
 
 on:
   schedule:
