+++ scripts/ci/validate_maven_publication.py
@@ -22,7 +22,7 @@ PUBLIC_ARTIFACTS = {
     "shaft-browserstack": (Path("shaft-browserstack/pom.xml"), "jar"),
     "shaft-video": (Path("shaft-video/pom.xml"), "jar"),
     "shaft-visual": (Path("shaft-visual/pom.xml"), "jar"),
-    "SHAFT_MCP": (Path("shaft-mcp/pom.xml"), "jar"),
+    "shaft-mcp": (Path("shaft-mcp/pom.xml"), "jar"),
     "shaft-bom": (Path("shaft-bom/pom.xml"), "pom"),
     "SHAFT_ENGINE": (Path("legacy-shaft-engine/pom.xml"), "pom"),
 }
