+++ scripts/ci/validate_quality_configuration.py
@@ -11,7 +11,7 @@ ROOT = Path(__file__).resolve().parents[2]
 NS = {"m": "http://maven.apache.org/POM/4.0.0"}
 JAVA_MODULES = {
     "shaft-engine", "shaft-pilot-core", "shaft-capture", "shaft-doctor", "shaft-ai", "shaft-heal",
-    "shaft-browserstack", "shaft-video", "shaft-visual", "SHAFT_MCP",
+    "shaft-browserstack", "shaft-video", "shaft-visual", "shaft-mcp",
 }
 DEPENDABOT_DIRECTORIES = {
     "/",
