+++ scripts/ci/verify_maven_central_release.py
@@ -18,7 +18,7 @@ GROUP_PATH = "io/github/shafthq"
 DEFAULT_REPOSITORY = "https://repo.maven.apache.org/maven2"
 JAR_ARTIFACTS = (
     "shaft-engine", "shaft-pilot-core", "shaft-capture", "shaft-doctor", "shaft-ai", "shaft-heal",
-    "shaft-browserstack", "shaft-video", "shaft-visual", "SHAFT_MCP",
+    "shaft-browserstack", "shaft-video", "shaft-visual", "shaft-mcp",
 )
 POM_ARTIFACTS = ("shaft-parent", "shaft-bom", "SHAFT_ENGINE")
 FIXTURE_GOALS = {
