+++ shaft-mcp/pom.xml
@@ -10,9 +10,9 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>SHAFT_MCP</artifactId>
+    <artifactId>shaft-mcp</artifactId>
     <packaging>jar</packaging>
-    <name>SHAFT_MCP</name>
+    <name>shaft-mcp</name>
     <description>Model Context Protocol server for SHAFT browser automation.</description>
 
     <properties>
@@ -124,7 +124,7 @@
                 <artifactId>maven-jar-plugin</artifactId>
                 <configuration>
                     <includes>
-                        <include>io/**</include>
+                        <include>com/**</include>
                         <include>application*.properties</include>
                         <include>logback-spring.xml</include>
                     </includes>
@@ -135,7 +135,7 @@
                 <artifactId>maven-source-plugin</artifactId>
                 <configuration>
                     <includes>
-                        <include>io/**</include>
+                        <include>com/**</include>
                     </includes>
                 </configuration>
             </plugin>
@@ -143,8 +143,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-javadoc-plugin</artifactId>
                 <configuration>
-                    <header>SHAFT_MCP, ${project.version}</header>
-                    <doctitle>SHAFT_MCP, ${project.version}</doctitle>
+                    <header>shaft-mcp, ${project.version}</header>
+                    <doctitle>shaft-mcp, ${project.version}</doctitle>
                 </configuration>
             </plugin>
             <plugin>
