+++ shaft-mcp/server.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://static.modelcontextprotocol.io/schemas/2025-10-17/server.schema.json",
   "name": "io.github.ShaftHQ/shaft-mcp",
-  "title": "SHAFT MCP",
+  "title": "shaft-mcp",
   "description": "SHAFT browser automation, capture, and offline deterministic failure diagnosis",
   "version": "@project.version@",
   "repository": {
