+++ shaft-mcp/src/main/java/com/shaft/mcp/BrowserService.java
@@ -1,4 +1,4 @@
-package io.github.shafthq.SHAFT_MCP;
+package com.shaft.mcp;
 
 import com.shaft.driver.SHAFT;
 import org.slf4j.Logger;
@@ -6,7 +6,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.ai.tool.annotation.Tool;
 import org.springframework.stereotype.Service;
 
-import static io.github.shafthq.SHAFT_MCP.EngineService.getDriver;
+import static com.shaft.mcp.EngineService.getDriver;
 
 @Service
 public class BrowserService {
