+++ shaft-mcp/src/main/resources/application-http.properties
@@ -1,5 +1,5 @@
 # Streamable HTTP transport configuration for hosted deployments
-spring.application.name=SHAFT_MCP
+spring.application.name=shaft-mcp
 spring.main.web-application-type=servlet
 spring.ai.mcp.server.name=shaft-mcp
 spring.ai.mcp.server.version=@project.version@
@@ -16,5 +16,5 @@ spring.ai.mcp.server.sse-endpoint=/mcp
 # Enable banner and logging for HTTP mode
 spring.main.banner-mode=console
 logging.level.root=INFO
-logging.level.io.github.shafthq.SHAFT_MCP=DEBUG
+logging.level.com.shaft.mcp=DEBUG
 logging.pattern.console=%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n
