+++ shaft-mcp/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=SHAFT_MCP
+spring.application.name=shaft-mcp
 spring.main.web-application-type=none
 spring.ai.mcp.server.name=shaft-mcp
 spring.ai.mcp.server.version=@project.version@
