+++ shaft-mcp/src/test/resources/fixtures/shaft-pilot/mcp/claude-desktop.json
@@ -2,7 +2,7 @@
   "mcpServers": {
     "shaft-mcp": {
       "command": "java",
-      "args": ["-jar", "/absolute/path/SHAFT_MCP-<version>.jar"]
+      "args": ["-jar", "/absolute/path/shaft-mcp-<version>.jar"]
     }
   }
 }
