+++ shaft-mcp/src/test/resources/fixtures/shaft-pilot/mcp/codex-config.toml
@@ -1,4 +1,4 @@
 [mcp_servers.shaft-mcp]
 command = "java"
-args = ["-jar", "/absolute/path/SHAFT_MCP-<version>.jar"]
+args = ["-jar", "/absolute/path/shaft-mcp-<version>.jar"]
 default_tools_approval_mode = "prompt"
