+++ shaft-mcp/src/test/resources/fixtures/shaft-pilot/mcp/vscode-mcp.json
@@ -3,7 +3,7 @@
     "shaft-mcp": {
       "type": "stdio",
       "command": "java",
-      "args": ["-jar", "/absolute/path/SHAFT_MCP-<version>.jar"]
+      "args": ["-jar", "/absolute/path/shaft-mcp-<version>.jar"]
     }
   }
 }
